Moves files from combat to Campaign package
