+++ app/src/main/java/angelosz/catbattlegame/ui/campaign/EnemyType.kt
@@ -1,4 +1,4 @@
-package angelosz.catbattlegame.ui.combat
+package angelosz.catbattlegame.ui.campaign
 
 enum class EnemyType(){
     SIMPLE_ENEMY
